dnscheck/requirements.txt
dataclasses-json
docker
paramiko
dnspython